extract al_syntax and al_types as sibling front-end crates
al_syntax is the syntax layer: span, token, scanner, parser, AST,
formatter, the diagnostic machinery, and module *identity*
(ModulePath/ModuleKey/ResolveError, moved from al_core::module - a
Diagnostic carries the key of the module it points into). It depends on
nothing in the workspace, so it compiles in parallel with al_vm and is
reusable by tooling that never compiles a program.

al_types is the type layer: the HM inference engine, type_def,
exhaustiveness, and the generic labelled-slot matcher (slots.rs, moved
from typed_ir - the typechecker and elaborator share it). It depends on
al_syntax (it types the AST) and al_vm (TypeId, Op for @vm intrinsic
schemes), not on the compiler.

Three edges were cut to make the split, each one item: EntityKind moves
from reference into al_types (reference re-exports it), slot_labeled's
file moves whole, and ModuleKey's minting constructors go public for
al_core's resolver and reference interner (doc updated to name them).
Module *resolution* - the table, caching, imports - stays in
al_core::module, built on the moved identity types.

al_core re-exports both crates at the historical paths (al_core::parser,
al_core::types, ...), so the compiler, driver, and build.rs are unchanged.

The workspace is now five crates: al_vm (runtime) parallel to
al_syntax -> al_types, both under al_core (compiler), under al (driver).
