+++ crates/al_core/src/lib.rs
@@ -15,27 +15,22 @@
 // runtime.
 #![deny(unsafe_code)]
 
-pub mod ast;
 pub mod bytecode;
 pub mod core_ir;
-pub mod diagnostic;
-pub mod formatter;
 pub mod module;
-pub mod parser;
 pub mod precompile;
 pub mod reference;
-pub mod scanner;
-pub mod span;
 pub mod static_ir;
-pub mod term;
-pub mod token;
-pub mod type_def;
 pub mod typed_ir;
-pub mod types;
 
-// The runtime substrate — values, heap, frozen area, index vectors — lives in
-// `al_vm`; re-exported at the historical paths so `al_core::heap` etc. keep
-// naming the one shared definition.
+// The syntax layer (source → AST, diagnostics, formatting, module identity),
+// the type layer (HM inference, type definitions, exhaustiveness), and the
+// runtime substrate (values, heap, frozen area, index vectors) live in their
+// own crates; re-exported at the historical paths so `al_core::parser`,
+// `al_core::types`, `al_core::heap` etc. keep naming the one shared
+// definition.
+pub use al_syntax::{ast, diagnostic, formatter, parser, scanner, span, term, token};
+pub use al_types::{type_def, types};
 pub use al_vm::{assert_send, assert_send_sync, frozen, heap, tivec};
 
 pub use bytecode::{CtorRef, PreludeBindings, TypeRef};
