+++ crates/al_types/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "al_types"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+al_syntax = { path = "../al_syntax" }
+al_vm = { path = "../al_vm" }
+indexmap = { workspace = true }
+smallvec = "1.15"
